refactor: hrnet.py - hrnet의 원시출력으로 부터 원하는 부위만 히트맵을 생성하기 위한 일환으로 원시출력과 히트맵 생성에 대한 부분을 분리 - 히트맵 생성은 원시출력으로 부터 한사람(객체) 분의 히트맵을 생성하는 get_heatmaps 함수와 한 프레임 내에 있는 모든 히트맵을 이미지 크기로 합치는 merge_heatmaps 함수로 분리 - get_heatmaps 함수 호출 시 컬러맵을 지정해주면 3차원 어레이로, 지정하지 않으면 2차원 어레이로 리턴
