+++ utils/visualization.py
@@ -161,19 +161,41 @@ def imshow_keypoints(img,
     return img
 
 
+def get_heatmaps(batch_heatmaps, colormap=None):
+    heatmaps = []
+    for _heatmaps in batch_heatmaps:
+        new_heatmap = np.zeros((_heatmaps.shape[1], _heatmaps.shape[2]), dtype=np.float32)
+        for heatmap in _heatmaps:
+            new_heatmap = np.maximum(new_heatmap, heatmap)
+
+        # new_heatmap = cv2.resize(new_heatmap, [new_heatmap.shape[1], new_heatmap.shape[0]])
+        # print(new_heatmap.shape)
+        if colormap is not None:
+            new_heatmap = new_heatmap * 255
+            new_heatmap = cv2.applyColorMap(new_heatmap.astype(np.uint8), colormap)
+        heatmaps.append(new_heatmap)
+    return heatmaps
+
+
 def merge_heatmaps(heatmaps, boxes, img_size):
-    heatmap = np.zeros((img_size[0], img_size[1]), dtype=np.float32)
+    if len(heatmaps[0].shape) == 3:
+        heatmap = np.zeros((img_size[0], img_size[1], img_size[2]), dtype=np.float32)
+    else:
+        heatmap = np.zeros((img_size[0], img_size[1]), dtype=np.float32)
 
     for h, b in zip(heatmaps, boxes):
-        new_heatmap = np.zeros((img_size[0], img_size[1]), dtype=np.float32)
+        if len(heatmaps[0].shape) == 3:
+            new_heatmap = np.zeros((img_size[0], img_size[1], img_size[2]), dtype=np.float32)
+        else:
+            new_heatmap = np.zeros((img_size[0], img_size[1]), dtype=np.float32)
         x1, y1, x2, y2 = int(b[0]), int(b[1]), int(b[2]), int(b[3])
         box_w, box_h = x2 - x1, y2 - y1
 
         # h = cv2.resize(h, (h.shape[1]*2, h.shape[0]*2))
-        h0, w0 = h.shape
+        h0, w0 = h.shape[:2]
         h = h[int(h0 * 0.1):int(h0 * 0.9), int(w0 * 0.1):int(w0 * 0.9)]
 
-        h0, w0 = h.shape
+        h0, w0 = h.shape[:2]
         box_ratio = box_w / box_h
         nw = h0 * box_ratio
         k = int((w0 - nw) / 2)
